+++ packages/emoji/docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* breaking; EmojiProvider requires getter/setter for remembering tone selection
+* breaking; skin tone selection made in the picker is stored in the EmojiResource (issues closed: fs-1127)
 * feature; emojis with skin variations in the picker (issues closed: fs-1126)
 ## 18.4.2 (2017-07-10)